+++ layoutdb/fuzz/fuzz_targets/fuzz_target_1.rs
@@ -12,7 +12,7 @@ fuzz_target!(|data: &[u8]| {
         // messages during the test.
         let mut diags = Diags::new("fuzz_target_1",str_in, 0, false);
         if let Ok(ast) = Ast::new(str_in, &mut diags) {
-            if let Ok(ast_db) = AstDb::new(&mut diags, &ast) {
+            if let Ok(ast_db) = AstDb::new(&mut diags, &ast, true) {
                 let _ = LinearDb::new(&mut diags, &ast, &ast_db);
             }
         }