+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   The audit-check action on github requires the lockfile.
 - Added std::md5 extension.
 - Added a vscode syntax highlighting extension.  See the bottom of the README for instructions.
+- `if/else` expressions still require const conditional evaluation, but the
+  conditional blocks can now contain structural statements, e.g. wr, set_addr, etc.
 
 ## [5.0.5] - 2026-04-09